requirements.txt
sanic==23.6.0
pyjwt==2.8.0
sanic_session==0.8.0
pymongo==4.5.0
fastapi==0.104.1
uvicorn==0.23.2
email-validator==2.1.0.post1
passlib==1.7.4
python-jose==3.3.0
loguru==0.7.2